+++ js/data.js
@@ -1,0 +1,43 @@
+const DESCRIPTIONS = [
+  'Зимний отпуск',
+  'Прогулка в парке',
+  'День рождения',
+  'На свадьбе',
+  'В командировке',
+];
+
+const COMMENTS = [
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+  'Всё отлично!',
+  'Поняла, что давно не ходила по магазинам, когда отвалилась подошва от сандалей 😂 не будьте как я, не доводите свои любимые вещи до такого состояния ☺ в итоге нашла оч красивое платье 👗, крутость которого не передаёт мобильная фотография 😮 иногда фотографу нужен фотограф 😂',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Ну и как это комментировать?',
+  'В целом всё неплохо.',
+  'Шикарная!',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Какая лучшая работа в мире? Все верно. Моя. Ну что тут сказать, #обожаю 😂',
+  'Какой глубокий хороший взгляд...',
+  'В целом всё неплохо. Хорошо бы убирать палец из кадра.',
+  'Как на парад.',
+  'Деньги на ветер. Это непрофессионально.',
+  'Уиии!!!!!!',
+  'Всё видно! Но не всё отлично! Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+  'По крайней мере все остались живы.',
+  'В целом всё неплохо. Но лица у людей на фотке перекошены, как будто их избивают.',
+  'Раньше было лучше!',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. Как можно было поймать такой неудачный момент?!',
+  'Терпимо. Руки есть.',
+  'В конце концов это просто непрофессионально. Поскользнулся на банановой кожуре и уронил фотоаппарат на кота.',
+  'Это просто непрофессионально.',
+  'Сойдёт для такой цены.',
+  'Всё отлично! Руки на месте.',
+  'Горизонт завален.',
+  'Друзья, пост благодарности! Мои клиенты - вы коты вообще котята!!! Тогда, когда мне некогда не то что выкладывать посты, но и просто вздохнуть, вы выкладываете свои фотографии, сделанные мной, да ещё и подписываете красиво-красиво, пишите просто волшебные слова. Хоть я не всегда успеваю следить, но когда читаю - прихожу в восторг! Спасибо, что вы такие крутянские ❤❤❤',
+  'Я сегодня буду спамить 😂 я НЕ следующая, такая табличка досталась 😂',
+  'Старая дева в окошке.. подарите ей кошку 🐈 😆Последние два дня я просто #счастливкакникто',
+  'Запретили выкладывать фотки со вчерашнего мероприятия, поэтому ловите утреннюю меня 😂 всем хорошего воскресенья ☺',
+];
+
+export {DESCRIPTIONS, COMMENTS};
